Added an assert into the trace reader to ensure the opcode address is different than zero.

+++ utils.cpp
@@ -55,8 +55,8 @@ uint64_t utils_t::get_power_of_two(uint64_t n) {
 };
 
 // =============================================================================
-uint8_t utils_t::check_if_power_of_two(uint64_t n) {
-    uint8_t i, count = 0;
+uint32_t utils_t::check_if_power_of_two(uint64_t n) {
+    uint32_t i, count = 0;
     uint64_t pow;
     for (i = 0; i < 64 && n != 0 ; i++) {
         pow = n & (1 << i);
